Implementing starRating per user Code cleanup

+++ src/main/java/be/syntra/devshop/DevshopFront/models/StarRating.java
@@ -19,5 +19,5 @@ public class StarRating {
     private String userName;
 
     @PositiveOrZero
-    private int rating;
+    private double rating;
 }
